Added DirkSimple.continue_clip.
Mad Dog McCree's tutorial level has footage of glass bottles on a fence,
shattering as they are shot. It shows a single frame of the bottle, and if
the player shoots it, it starts the laserdisc playing again, animating the
destruction, so it looks like it was a solid bottle that you just put a bullet
through.

This functionality allows that to work smoothly, without having to use a seek
in the Theora file to accomplish this effect.