maddog: Initial work on Mad Dog McCree.
Been working on this awhile; it's not close to done, but it's irresponsible
to not put it in revision control at this point.

This is based off a Singe 2 implementation of Mad Dog McCree HD, where I've
taken some sounds and graphics, and benefitted from having a clear
description of the game logic and scene timings, but the code for DirkSimple
was written from scratch, and operates in a dramatically different way.

This implementation is currently enough to see the attract mode, play the
tutorial and first level, and deal with deaths/continues/game over. An
enormous amount of infrastructure is in place now, so most of the remaining
work is just filling in the rest of the actual game scenes (at least, I hope
so).