theoraplay: Reworked seeking.
This is now frame-perfect, afaict, which fixes a _lot_ of things: the bottles
in the maddog tutorial don't start broken, single-frame things like the
town crossroads level select show the right image, audio isn't breaking up
anymore, etc.

The problems though: seeks are noticibly slower now...and there's a ton of
things I adjusted in Dragon's Lair to be a frame or two different to work
around the imperfect seeking that probably need to be adjusted back now.

But this is still progress! I just need to keep working at it.