render: Allow clear_screen render commands to be for a single frame.
Every render command from Lua code is queued and rerun every frame, so that
UI stays consistent when the laserdisc is seeking and the Lua tick isn't run.

But Mad Dog McCree uses a screen clear to flash the screen when firing the
gun, and we don't want the whole screen to stay white during the duration of
the seek, so allow these to be marked as only meant for a single-frame of
rendering.

This is slightly goofy, but the alternative is either rejigger the Lua tick to
run during seeks in every game, or add a second optional tick_ui_only function
that runs during seeks (probably the more correct approach, but also the more
complicated).