theoraplay: Changed seeking strategy.
This is back to the less-correct but much faster approach, but I added in the
seek_only_backwards trick, which was solving the accuracy problem; eventually
the bisecting would choose the same seek position it just tried and would have
to give up; at this point, now we start moving backwards exclusively until
we have a reasonable position and call it a day.

Still fast, but with all the bug fixes!